+++ UCR.Core/Context.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
@@ -31,10 +31,12 @@ namespace HidWizards.UCR.Core
         [XmlIgnore] public SubscriptionsManager SubscriptionsManager { get; set; }
         [XmlIgnore] public PluginsManager PluginManager { get; set; }
         [XmlIgnore] public BindingManager BindingManager { get; set; }
+        [XmlIgnore] public SettingsManager SettingsManager { get; set; }
 
         public delegate void ActiveProfileChanged(Profile profile);
         public event ActiveProfileChanged ActiveProfileChangedEvent;
-        
+        public event Action MinimizedToTrayEvent;
+
         internal bool IsNotSaved { get; private set; }
         internal IOController IOController { get; set; }
         private OptionSet options;
@@ -59,6 +61,7 @@ namespace HidWizards.UCR.Core
                 Logger.Error("IOWrapper provider directory not found", e);
             }
             
+            SettingsManager = new SettingsManager();
             ProfilesManager = new ProfilesManager(this, Profiles);
             DevicesManager = new DevicesManager(this);
             SubscriptionsManager = new SubscriptionsManager(this);
@@ -69,8 +72,14 @@ namespace HidWizards.UCR.Core
         private void SetCommandLineOptions()
         {
             options = new OptionSet {
-                { "p|profile=", "The profile to search for", FindAndLoadProfile }
+                { "p|profile=", "The profile to search for", FindAndLoadProfile },
+                { "h|hidden", "Minimize to system tray", x => MinimizeToTray() }
             };
+        }
+        
+        public void MinimizeToTray()
+        {
+            MinimizedToTrayEvent.Invoke();
         }
 
         private void FindAndLoadProfile(string profileString)
@@ -98,7 +107,7 @@ namespace HidWizards.UCR.Core
         }
 
         #region Persistence
-        
+
         public bool SaveContext(List<Type> pluginTypes = null)
         {
             var serializer = GetXmlSerializer(pluginTypes);