+++ UCR/App.xaml.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Runtime.InteropServices;
 using System.Windows;
 using HidWizards.UCR.Core;
+using HidWizards.UCR.Core.Models.Settings;
 using HidWizards.UCR.Core.Utilities;
 using HidWizards.UCR.Utilities;
 using HidWizards.UCR.Views;
@@ -19,6 +20,7 @@ namespace HidWizards.UCR
         private Context context;
         private HidGuardianClient _hidGuardianClient;
         private SingleGlobalInstance mutex;
+        private bool StartMinimized;
 
         protected override void OnStartup(StartupEventArgs e)
         {
@@ -35,15 +37,22 @@ namespace HidWizards.UCR
                 InitializeUcr();
                 CheckForBlockedDll();
 
-                context.ParseCommandLineArguments(e.Args);
                 var mw = new MainWindow(context);
-                mw.Show();
+                context.MinimizedToTrayEvent += Context_MinimizedToTrayEvent;
+                context.ParseCommandLineArguments(e.Args);
+                if (!StartMinimized && SettingsCollection.LaunchMinimized) context.MinimizeToTray();
+                if (!StartMinimized) mw.Show();
             }
             else
             {
                 SendArgs(string.Join(";", e.Args));
                 Current.Shutdown();
             }
+        }
+
+        private void Context_MinimizedToTrayEvent()
+        {
+            StartMinimized = true;
         }
 
         private void InitializeUcr()