+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+- Added Settings
+- New setting: Start minimized
+- New setting: Minimize to tray on close
+- Added System Tray Icon
+- Added Hidden mode CLI Argument (-h|hidden)
+
 ## [0.9.0] - 2020-01-02
 
 ### Added